Move model in folder Models
