+++ client/app/comment/comment-list.component.ts
@@ -1,7 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 
 import {CommentService} from "./comment.service";
-import {Comment} from "./comment.model";
+import {Comment} from "../models/comment.model";
 
 @Component({
     selector: "comment-list",
