+++ client/app/models/comment.model.ts
